chunk cards when importing csv
